requirements.txt
# gRPC and Protocol Buffers
grpcio
grpcio-tools
protobuf

# Database
SQLAlchemy
psycopg2-binary
alembic

# Authentication and Security
PyJWT
bcrypt
cryptography

# Email
python-dotenv

# Utilities
python-dateutil
